perf(widescreen): M1.2 explicit uint32_t at window call sites
Cast the int width/height members to uint32_t at every
window->SetParameter(PARAM_WIDTH/HEIGHT, ...) call site so the stored
variant alternative matches what WindowNode::GetParameterValue<uint32_t>
expects, rather than relying solely on M1.1's runtime conversion.
Fixed all five graph builders that configure a window node:
BuildRenderGraph, BuildInstancingDemoGraph, BuildAutoSyncDemoGraph,
BuildUIGraph, BuildFanInDemoGraph.

Audited every other integral GetParameterValue<uint32_t>/<int> reader
against its SetParameter call sites (BenchmarkGraphFactory.cpp and the
rest of RenderGraph's nodes) — all already store matching or explicitly
cast types, so no further call-site changes were needed.

Live lavapipe gate at 1280x720 confirms the fix: dispatch dims compute
as 160x90x1 (from window 1280x720) instead of the old 800x600-derived
100x75, and the perf summary reports Resolution: 1280x720. No
NodeParameterManager type-mismatch warnings fired during the run.

+++ VIXEN/application/main/source/graph/BuildFanInDemoGraph.cpp
@@ -147,8 +147,8 @@ void VulkanGraphApplication::BuildFanInDemoGraph() {
     // Parameters
     // ===================================================================
     auto* window = static_cast<WindowNode*>(renderGraph->GetInstance(windowNode));
-    window->SetParameter(WindowNodeConfig::PARAM_WIDTH, width);
-    window->SetParameter(WindowNodeConfig::PARAM_HEIGHT, height);
+    window->SetParameter(WindowNodeConfig::PARAM_WIDTH, static_cast<uint32_t>(width));
+    window->SetParameter(WindowNodeConfig::PARAM_HEIGHT, static_cast<uint32_t>(height));
     auto* device = static_cast<DeviceNode*>(renderGraph->GetInstance(deviceNode));
     device->SetParameter(DeviceNodeConfig::PARAM_GPU_INDEX, 0u);
