+++ VIXEN/application/main/source/graph/BuildUIGraph.cpp
@@ -44,8 +44,8 @@ void VulkanGraphApplication::BuildUIGraph() {
     NodeHandle presentNode     = renderGraph->AddNode<PresentNodeType>("ui_present");
 
     auto* window = static_cast<WindowNode*>(renderGraph->GetInstance(windowNode));
-    window->SetParameter(WindowNodeConfig::PARAM_WIDTH, width);
-    window->SetParameter(WindowNodeConfig::PARAM_HEIGHT, height);
+    window->SetParameter(WindowNodeConfig::PARAM_WIDTH, static_cast<uint32_t>(width));
+    window->SetParameter(WindowNodeConfig::PARAM_HEIGHT, static_cast<uint32_t>(height));
     auto* device = static_cast<DeviceNode*>(renderGraph->GetInstance(deviceNode));
     device->SetParameter(DeviceNodeConfig::PARAM_GPU_INDEX, 0u);
     auto* present = static_cast<PresentNode*>(renderGraph->GetInstance(presentNode));
